refactor(tab-buttons): internalize Pill primitive
