--- src/components/Pill/Pill.api.md
@@ -1,99 +1,0 @@
-<!-- Auto Generated by scripts/propsgen.ts -->
-<script setup>
-  import PropsTable from '@/components/Docs/PropsTable.vue'
-  import SlotsTable from '@/components/Docs/SlotsTable.vue'
-  import EmitsTable from '@/components/Docs/EmitsTable.vue'
-
-  const propsData = [
-  {
-    name: 'label',
-    description: 'Text shown inside the pill.',
-    required: false,
-    type: 'string | number'
-  },
-  {
-    name: 'variant',
-    description: 'Visual variant.',
-    required: false,
-    type: 'PillVariant',
-    default: '"default"'
-  },
-  {
-    name: 'size',
-    description: 'Size of the pill.',
-    required: false,
-    type: 'PillSize',
-    default: '"md"'
-  },
-  {
-    name: 'active',
-    description: 'Active/selected state.',
-    required: false,
-    type: 'boolean',
-    default: 'false'
-  },
-  {
-    name: 'icon',
-    description: 'Icon-only content. When set, the pill renders as a square icon button and any `label` is exposed only to assistive tech (sr-only).',
-    required: false,
-    type: 'PillIcon'
-  },
-  {
-    name: 'iconLeft',
-    description: 'Leading accent icon, rendered next to a visible label.',
-    required: false,
-    type: 'PillIcon'
-  },
-  {
-    name: 'iconRight',
-    description: 'Trailing accent icon, rendered next to a visible label.',
-    required: false,
-    type: 'PillIcon'
-  },
-  {
-    name: 'browserTabBase',
-    description: 'Browser-tab active edge shape.',
-    required: false,
-    type: 'BrowserTabBase',
-    default: '"none"'
-  },
-  {
-    name: 'orientation',
-    description: 'Axis used by underline tabs.',
-    required: false,
-    type: 'PillOrientation',
-    default: '"horizontal"'
-  },
-  {
-    name: 'activeStyle',
-    description: 'Active treatment for default pills.',
-    required: false,
-    type: 'PillActiveStyle',
-    default: '"raised"'
-  }
-]
-
-  const slotsData = [
-  {
-    name: 'prefix',
-    description: '',
-    type: 'any'
-  },
-  {
-    name: 'default',
-    description: '',
-    type: 'any'
-  },
-  {
-    name: 'suffix',
-    description: '',
-    type: 'any'
-  }
-]
-</script>
-## API Reference
-
-<PropsTable name="Pill" :data="propsData"/> 
-
-<SlotsTable :data="slotsData"/> 
-
