--- src/components/Pill/Pill.md
@@ -1,20 +1,0 @@
-# Pill
-
-A compact tab item primitive used by segmented tabs, underline tabs, and
-browser-style tabs.
-
-## Playground
-
-<ClientOnly><PillBuilder /></ClientOnly>
-
-## Variants
-
-<ComponentPreview name="Pill-Variants" />
-
-## Sizes
-
-<ComponentPreview name="Pill-Sizes" />
-
-## Browser edges
-
-<ComponentPreview name="Pill-BrowserEdges" />
