--- src/components/Pill/index.ts
@@ -1,10 +1,0 @@
-export { default as Pill } from './Pill.vue'
-export type {
-  BrowserTabBase,
-  PillActiveStyle,
-  PillIcon,
-  PillOrientation,
-  PillProps,
-  PillSize,
-  PillVariant,
-} from './types'
