+++ src/components/TabButtons/TabButtons.vue
@@ -2,7 +2,8 @@
 import { computed, watch, watchEffect, h, type FunctionalComponent } from 'vue'
 import { RadioGroupItem, RadioGroupRoot } from 'reka-ui'
 import { RouterLink } from 'vue-router'
-import { Pill, type BrowserTabBase } from '../Pill'
+import Pill from './Pill.vue'
+import type { BrowserTabBase } from './pillTypes'
 import { warnDeprecated } from '../../utils/warnDeprecated'
 import { warnFeatherIconUsage } from '../../utils/iconString'
 import type { TabButton, TabButtonsEmits, TabButtonsProps } from './types'
