+++ src/index.ts
@@ -49,7 +49,6 @@ export * from './components/MonthPicker'
 export * from './components/MultiSelect'
 export * from './components/Password'
 export * from './components/Popover'
-export * from './components/Pill'
 export * from './components/Progress'
 export * from './components/Rating'
 export { default as Resource } from './components/Resource.vue'
